requirements.txt
numpy
jupyterlab
scikit-learn
openai
mem0ai
